Move AF callbacks to server-specific directory
