+++ packages/lesswrong/server.js
@@ -92,6 +92,9 @@ import './server/callbacks/reviewVoteCallbacks';
 if (getSetting('hasEvents', true)) {
   import './server/callbacks/localgroupCallbacks';
 }
+import './server/callbacks/alignment-forum/alignmentCommentCallbacks';
+import './server/callbacks/alignment-forum/alignmentPostCallbacks';
+import './server/callbacks/alignment-forum/alignmentUserCallbacks';
 
 import './lib/collections/revisions/resolvers.js';
 import './lib/collections/posts/serverSchema.js';
